feature/CLA-006-get-claim
This MR covers get feature being added to the claim system.
## Why?
This feature is being implemented to meet read (get) requirements.
## How?
Controller has had a method 'getClaimById' added which takes a claim ID number, passes it to ClaimService.getClaimById, and returns a response depending on the context of the ID. Claim Service uses an implementation to check the database for the given ID and feeds a response back to the controller. If the claim does not exist, an exception is thrown. If it is, the service will map the found  claim to a Dto and return it to the controller.
This exception has been added in the form of a ResourceNotFoundException, which extends a RuntimeException. As inferred, this exception will be thrown when the resource is not found via the given ID.
## Testing?
Unit tests have been added for the controller to test that the finding of a claim by ID was called once with the correct ID, and that the response expected matches the actual result. A check for the Dto being null is also checked just to make sure that no writing is happening at this step.
Unit tests for the ClaimService have also been added to make sure that the actual outcomes are happening as expected when the repository is given a claim and when the repository does not have a claim.