CLA-007-get-all-claims
This MR covers get all (read all) feature being added to the claim system.
## Why?
This feature is being implemented to meet read (get) requirements.
## How?
Controller has had a method 'getAllClaims' added which creates a http request for all claims at the endpoint api/claims. Claim Service uses an implementation to check the database for all claims and returns them as a list to the controller.
## Testing?
Unit tests have been added for the controller to test that the finding of all claims was successfully, and that the response expected matches the actual result.
Unit tests for the ClaimService have also been added to make sure that the actual outcomes are happening as expected when the repository is given multiple claims.