CLA-008-update-claim
This MR covers update functionality being added to the claim system.
## Why?
This feature is being implemented to meet update requirements as part of the CRUD assignment.
## How?
PutMapping added to ClaimController to find an entry in the database via id.
Service and implementation added to update the found entry as long as the entry exists and the updated date is not before created date. If the dates do not fit the criteria, an exception has been added to throw when this is the case.
## Testing?
Unit Test Coverage added for all methods.