CLA-011-note-repository
## What?
This MR covers note repository being added to the claim system.
## Why?
As part of the CRUD assignment, Note repository is needed alongside claim. Repositories are an important component of CRUD functionality, as they provide the mechanisms for CRUD operations on objects.
## How?
Note Repository interface has been created and extends JpaRepository.
## Testing?
Unit test coverage will be included as functionality is added for note.