CLA-012-note-dto-and-mapper
## What?
This MR covers note dto and note mapper being added to the claim system.
## Why?
As part of the CRUD assignment, Note dto and mapper is needed fo CRUD functionality. DTOs are in place to have data from a database mapped to them.
## How?
NoteMapper and NoteDto classes have been created and populated with necessary code for functionality.
## Testing?
Unit test coverage will be included as functionality is added for note.