CLA-013-note-rest-api-create-functionality
## What?
This MR covers note CRUD create functionality.
## Why?
C in CRUD is create. This operation is needed for one of the four essential operations for this assignment.
## How?
Controller has been added. Service and Implementation has been added. The controller creates a postmapping request for the service to create a new note. The service includes functionality to add, or create a note and save it to the database.
## Testing?
Unit test coverage has been added as part of this MR for Controller, Mapper and Service.