CLA-018-non-update-created-time (#18)
## What?
This MR is for note update functionality.
## Why?
As part of system's requirements, there would be no use case where the created time would need updating.
## How?
This has been amended so that created time will not be written over when doing a put request at an existing id. Furthermore, null checks have been added for each field that can be updated so that if a value is not passed, then the existing record will not write over these values with null. This has been added for claims and notes.
## Testing?
Unit test coverage not implemented for null scenarios as per this MR.